// Captioneer.UI/src/app/models/movie-viewmodel.ts
export class MovieViewModel {
    title: string = "";
    imdbId: string = "";
    synopsis: string = "";
    year: string = "";
    runtime?: number;
    imdbRatingValue?: number;
    imdbRatingCount?: number;
    rottenTomatoesValue: string = "";
    metacriticValue: string = "";
    coverArt: string = "";
}